requirements.txt
alembic==0.9.1
appdirs==1.4.3
Chameleon==3.1
click==6.7
croniter==0.3.16
dogpile.cache==0.6.2
Flask==0.12
Flask-Admin==1.5.0
hupper==0.4.4
itsdangerous==0.24
Jinja2==2.9.5
jsl==0.2.4
jsonschema==2.6.0
Mako==1.0.6
MarkupSafe==1.0
mock==2.0.0
packaging==16.8
PasteDeploy==1.5.2
pbr==2.0.0
psycopg2==2.7.1
Pygments==2.2.0
pymemcache==1.4.2
pyparsing==2.2.0
pyramid==1.9.1
pyramid-chameleon==0.3
pyramid-debugtoolbar==3.0.5
pyramid-dogpile-cache==0.0.4
pyramid-jinja2==2.6.2
pyramid-mako==1.0.2
pyramid_swagger_spec==0.0.7
pyramid-tm==1.1.1
python-crontab==2.2.2
python-dateutil==2.5.3
python-editor==1.0.3
pytz==2016.10
raven==6.0.0
repoze.lru==0.6
six==1.10.0
SQLAlchemy==1.1.7
transaction==2.1.2
translationstring==1.3
venusian==1.0
waitress==1.0.2
WebOb==1.7.2
Werkzeug==0.12.1
WTForms==2.1
zope.deprecation==4.2.0
zope.interface==4.4.2
zope.sqlalchemy==0.7.7
